% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{bam_check_nas}
\alias{bam_check_nas}
\title{Add missing-data inputs to data list}
\usage{
bam_check_nas(datalist)
}
\arguments{
\item{datalist}{a list of BAM inputs}
}
\description{
Binary matrices indicating where data are/aren't missing are
added to the data list. This is required in order to run
ragged-array data structures in the stanfile.
}
\details{
Previously this function omitted any times with missing data,
but now that ragged arrays are accommodated in the stanfile the
operations are entirely different.
}
