% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{bam_hydrograph}
\alias{bam_hydrograph}
\title{Plot flow time series from BAM inference}
\usage{
bam_hydrograph(fit, qobs = NULL)
}
\arguments{
\item{fit}{A stanfit object, as returned from \code{bam_estimate()}}

\item{qobs}{An optional vector giving observed flow for comparison}
}
\description{
Plot flow time series from BAM inference
}
