% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{bam_plot.bamdata}
\alias{bam_plot.bamdata}
\title{Plot a bamdata object}
\usage{
\method{bam_plot}{bamdata}(bamobj)
}
\arguments{
\item{bamobj}{A bamdata object, as produced by \code{bam_data()}}
}
\description{
Plot a bamdata object
}
