% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{bam_plot.bamval}
\alias{bam_plot.bamval}
\title{Plot a bamval object to show predictive performance}
\usage{
\method{bam_plot}{bamval}(bamobj)
}
\arguments{
\item{bamobj}{a \code{bamval} object, as produced by \code{bam_validate}}
}
\description{
Plot a bamval object to show predictive performance
}
