% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc.R
\name{bam_qpred}
\alias{bam_qpred}
\title{Flow posterior mean and Bayesian credible interval.}
\usage{
bam_qpred(fit, chain = "all", conf.level = 0.95)
}
\arguments{
\item{fit}{A stanfit object, as returned from \code{bam_estimate()}}

\item{chain}{Either an integer specifying which chain(s) to extract statistics from,
or "all" (the default), in which case they are extracted from all chains.}

\item{conf.level}{A numeric value on (0, 1) specifying the size of the Bayesian 
credible interval. Default is 0.95.}
}
\description{
Flow posterior mean and Bayesian credible interval.
}
