% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_settings.R
\name{bam_settings_unsupervised}
\alias{bam_settings_unsupervised}
\title{Options manager for geoBAMr defaults using unsupervised classification}
\usage{
bam_settings_unsupervised(..., .__defaults = FALSE, .__reset = FALSE)
}
\arguments{
\item{...}{(Optional) named settings to query or set.}

\item{.__defaults}{See \code{?settings::option_manager}}

\item{.__reset}{See \code{?settings::option_manager}}
}
\description{
Options manager for geoBAMr defaults using unsupervised classification
}
