% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{bam_valdata}
\alias{bam_valdata}
\title{Create a data.frame for BAM validation}
\usage{
bam_valdata(fit, qobs)
}
\arguments{
\item{fit}{A stanfit object, as returned from \code{bam_estimate()}}

\item{qobs}{a vector of observed flow.}
}
\description{
Create a data.frame for BAM validation
}
