% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{bam_validate}
\alias{bam_validate}
\title{Calculate validation metrics and plots}
\usage{
bam_validate(
  fit,
  qobs,
  stats = c("RRMSE", "MRR", "SDRR", "NSE", "NRMSE", "rBIAS", "CoV", "logNSE", "Ej")
)
}
\arguments{
\item{fit}{A stanfit object, as returned from \code{bam_estimate()}}

\item{qobs}{a vector of observed flow.}

\item{stats}{Which stats to include in the summary?}
}
\description{
Calculate validation metrics and plots
}
