% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify_func}
\alias{classify_func}
\title{Classify river for expert framework}
\usage{
classify_func(Wobs)
}
\arguments{
\item{Wobs}{observed widths matrix}
}
\description{
Classify river for expert framework
}
