% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify_func_unsupervised}
\alias{classify_func_unsupervised}
\title{Classify river for unsupervised framework}
\usage{
classify_func_unsupervised(Wobs)
}
\arguments{
\item{Wobs}{observed widths matrix}
}
\description{
Classify river for unsupervised framework
}
