% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{CoV}
\alias{CoV}
\title{Coefficient of variation}
\usage{
CoV(pred, meas)
}
\arguments{
\item{pred}{vector of predictions}

\item{meas}{vector of measurements}
}
\description{
Coefficient of variation
}
