% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv2sigma}
\alias{cv2sigma}
\title{Convert coefficient of variation to sigma parameter of lognormal diistribution}
\usage{
cv2sigma(cv)
}
\arguments{
\item{cv}{Coefficient of variation}
}
\description{
Convert coefficient of variation to sigma parameter of lognormal diistribution
}
