% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{Ej}
\alias{Ej}
\title{E_j general efficiency statistic from Criss and Winston (2008)}
\usage{
Ej(pred, meas, j = 1, bench = mean(meas))
}
\arguments{
\item{pred}{vector of predictions}

\item{meas}{vector of measurements}

\item{j}{j parameter from general efficiency statistic.}

\item{bench}{Some benchmark for comparison, defaults to the mean of meas.}
}
\description{
E_j general efficiency statistic from Criss and Winston (2008)
}
