% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_A0SD}
\alias{estimate_A0SD}
\title{Estimate base cross-sectional area SD using bam data}
\usage{
estimate_A0SD(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate base cross-sectional area SD using bam data
}
