% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_b}
\alias{estimate_b}
\title{Estimate AHG b exponent using bam data}
\usage{
estimate_b(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate AHG b exponent using bam data
}
