% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_bSD}
\alias{estimate_bSD}
\title{Estimate AHG b SD using bam data}
\usage{
estimate_bSD(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate AHG b SD using bam data
}
