% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logA0_unsupervised}
\alias{estimate_logA0_unsupervised}
\title{Estimate base cross-sectional area using bam data}
\usage{
estimate_logA0_unsupervised(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate base cross-sectional area using bam data
}
