% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logDbSD_unsupervised}
\alias{estimate_logDbSD_unsupervised}
\title{Estimate bankful depth SD using bam data}
\usage{
estimate_logDbSD_unsupervised(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate bankful depth SD using bam data
}
