% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logn}
\alias{estimate_logn}
\title{Estimate manning's n using bam data}
\usage{
estimate_logn(Sobs)
}
\arguments{
\item{Sobs}{Observed S, as a space-down, time-across matrix}
}
\description{
Estimate manning's n using bam data
}
