% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logn_unsupervised}
\alias{estimate_logn_unsupervised}
\title{Estimate manning's n using bam data}
\usage{
estimate_logn_unsupervised(Sobs, Wobs)
}
\arguments{
\item{Sobs}{Observed S, as a space-down, time-across matrix}

\item{Wobs}{Observed W, as a space-down, time-across matrix}
}
\description{
Estimate manning's n using bam data
}
