% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_lognSD_unsupervised}
\alias{estimate_lognSD_unsupervised}
\title{Estimate manning's n SD using bam data}
\usage{
estimate_lognSD_unsupervised(Wobs)
}
\arguments{
\item{Wobs}{Observed W, as a space-down, time-across matrix}
}
\description{
Estimate manning's n SD using bam data
}
