% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logr_unsupervised}
\alias{estimate_logr_unsupervised}
\title{Estimate channel shape using bam data}
\usage{
estimate_logr_unsupervised(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate channel shape using bam data
}
