% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logrSD}
\alias{estimate_logrSD}
\title{Estimate channel shape SD using bam data}
\usage{
estimate_logrSD(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate channel shape SD using bam data
}
