% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logWb}
\alias{estimate_logWb}
\title{Estimate bankful width using bam data}
\usage{
estimate_logWb(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate bankful width using bam data
}
