% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logWbSD}
\alias{estimate_logWbSD}
\title{Estimate bankful width SD using bam data}
\usage{
estimate_logWbSD(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate bankful width SD using bam data
}
