% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_lowerboundA0}
\alias{estimate_lowerboundA0}
\title{Estimate base cross-sectional area lowerbound using bam data}
\usage{
estimate_lowerboundA0(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate base cross-sectional area lowerbound using bam data
}
