% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_lowerboundb}
\alias{estimate_lowerboundb}
\title{Estimate AHG b lowerbound using bam data}
\usage{
estimate_lowerboundb(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate AHG b lowerbound using bam data
}
