% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_lowerboundlogDb}
\alias{estimate_lowerboundlogDb}
\title{Estimate bankful depth lower bound using bam data}
\usage{
estimate_lowerboundlogDb(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate bankful depth lower bound using bam data
}
