% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_lowerboundlogn}
\alias{estimate_lowerboundlogn}
\title{Estimate manning's n lowerbound using bam data}
\usage{
estimate_lowerboundlogn(Wobs)
}
\arguments{
\item{Wobs}{Observed W, as a space-down, time-across matrix}
}
\description{
Estimate manning's n lowerbound using bam data
}
