% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_lowerboundlogr}
\alias{estimate_lowerboundlogr}
\title{Estimate channel shape lowerbound using bam data}
\usage{
estimate_lowerboundlogr(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate channel shape lowerbound using bam data
}
