% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_lowerboundlogWb_unsupervised}
\alias{estimate_lowerboundlogWb_unsupervised}
\title{Estimate bankful width lower bound using bam data}
\usage{
estimate_lowerboundlogWb_unsupervised(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate bankful width lower bound using bam data
}
