% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_upperboundb_unsupervised}
\alias{estimate_upperboundb_unsupervised}
\title{Estimate AHG b upperbound using bam data}
\usage{
estimate_upperboundb_unsupervised(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate AHG b upperbound using bam data
}
