% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_upperboundlogDb_unsupervised}
\alias{estimate_upperboundlogDb_unsupervised}
\title{Estimate bankful depth upper bound using bam data}
\usage{
estimate_upperboundlogDb_unsupervised(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate bankful depth upper bound using bam data
}
