% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_upperboundlogn}
\alias{estimate_upperboundlogn}
\title{Estimate manning's n upperbound using bam data}
\usage{
estimate_upperboundlogn(Wobs)
}
\arguments{
\item{Wobs}{Observed W, as a space-down, time-across matrix}
}
\description{
Estimate manning's n upperbound using bam data
}
