% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_upperboundlogr_unsupervised}
\alias{estimate_upperboundlogr_unsupervised}
\title{Estimate channel shape upperbound using bam data}
\usage{
estimate_upperboundlogr_unsupervised(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate channel shape upperbound using bam data
}
