% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_upperboundlogWb}
\alias{estimate_upperboundlogWb}
\title{Estimate bankful width upper bound using bam data}
\usage{
estimate_upperboundlogWb(Wobs)
}
\arguments{
\item{Wobs}{Observed W,as a space-down, time-across matrix.}
}
\description{
Estimate bankful width upper bound using bam data
}
