% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoBAMr-package.R
\docType{package}
\name{geoBAMr-package}
\alias{geoBAMr-package}
\alias{geoBAMr}
\title{The 'geoBAMr' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
