% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{ln_sigsq}
\alias{ln_sigsq}
\title{Calculate lognormal sigma parameter based on truncated normal parameters}
\usage{
ln_sigsq(obs, err_sigma, a = 0)
}
\arguments{
\item{obs}{A numeric vector of observations}

\item{err_sigma}{Standard deviation of measurement error}

\item{a}{zero-reference for method of moments.}
}
\description{
Used to put measurement errors into original log-normal parameterization.
}
