% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{logNSE}
\alias{logNSE}
\title{NSE, computed on log-transformed residuals}
\usage{
logNSE(pred, meas)
}
\arguments{
\item{pred}{vector of predictions}

\item{meas}{vector of measurements}
}
\description{
NSE, computed on log-transformed residuals
}
