% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{NRMSE}
\alias{NRMSE}
\title{Normalized root-mean-square error}
\usage{
NRMSE(pred, meas)
}
\arguments{
\item{pred}{vector of predictions}

\item{meas}{vector of measurements}
}
\description{
Normalized root-mean-square error
}
