% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{NSE}
\alias{NSE}
\title{Nash-Sutcliffe efficiency}
\usage{
NSE(pred, meas)
}
\arguments{
\item{pred}{vector of predictions}

\item{meas}{vector of measurements}
}
\description{
Nash-Sutcliffe efficiency
}
