% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{RRMSE}
\alias{RRMSE}
\title{Relative root-mean-square error}
\usage{
RRMSE(pred, meas)
}
\arguments{
\item{pred}{vector of predictions}

\item{meas}{vector of measurements}
}
\description{
Relative root-mean-square error
}
