% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{SDRR}
\alias{SDRR}
\title{Standard deviation of relative residual}
\usage{
SDRR(pred, meas)
}
\arguments{
\item{pred}{vector of predictions}

\item{meas}{vector of measurements}
}
\description{
Standard deviation of relative residual
}
