% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{bam_plot}
\alias{bam_plot}
\title{Plot a geoBAM-created object}
\usage{
bam_plot(bamobj)
}
\arguments{
\item{bamobj}{Currently supported objects are bamdata objects (created by
\code{bam_data()}) and bamval objects (created by \code{bam_validate()}).}
}
\description{
Plot a geoBAM-created object
}
